Remove test suites
Tests suites are currently not really exposed or used by anything. With
the introduction of namespaces soon coupled with the ability to query by
labels, it further removes another use case for test suites.

This commit removes them for now. We can always bring them back if a
clearer use case presents itself.